pyproject.toml
[project]
name = "graphiti-fastmcp"
version = "1.0.1"
description = "Graphiti FastMCP Server"
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "fastmcp>=2.13.1",
    "openai>=1.91.0",
    "graphiti-core[falkordb]==0.24.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.0.0",
    "claude-agent-sdk>=0.1.10",
    "langchain-openai>=1.1.0",
    "langchain-qdrant>=1.1.0",
    "eunomia-mcp>=0.1.0",
]

[project.optional-dependencies]
azure = [
    "azure-identity>=1.21.0",
]
providers = [
    "google-genai>=1.8.0",
    "anthropic>=0.49.0",
    "groq>=0.2.0",
    "voyageai>=0.2.3",
    "sentence-transformers>=2.0.0",
]

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.10"
typeCheckingMode = "basic"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = ["E501"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
# Required by Pydantic on Python < 3.12
"typing.TypedDict".msg = "Use typing_extensions.TypedDict instead."

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[dependency-groups]
dev = [
    "faker>=37.12.0",
    "httpx>=0.28.1",
    "psutil>=7.1.2",
    "pyright>=1.1.404",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-timeout>=2.4.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.7.1",
]
